requirements.txt
streamlit
langchain
langchain-openai
openai
PyMuPDF
faiss-cpu
